pyproject.toml
[tool.poetry]
name = "applifting-exercise"
version = "0.1.0"
description = "Simple Product aggregator microservice created as a coding excercise"
authors = ["Filip Krul <gvarph006@gmail.com>"]
readme = "README.md"
packages = [{include = "applifting_exercise"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
httpx = "^0.24.1"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.18"
uvicorn = "^0.22.0"
psycopg2-binary = "^2.9.6"
psycopg2 = "^2.9.6"
pyjwt = "^2.7.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
colorlog = "^6.7.0"
pytest-watch = "^4.2.0"
pytest-mock = "^3.11.1"
passlib = "^1.7.4"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true

[tool.pytest.ini_options]
addopts = "--cov src --cov-report=lcov:coverage/lcov.info --cov-report=term "